refactor: groq 클라이언트 리펙터링
- apps/core/utils로 옮김
- 관련 함수와 테스트 코드 수정
- s3_presigned_url 관련 불필요한 파일 삭제
