+++ apps/qna/chatbot/__init__.py
@@ -1,11 +1,8 @@
-from .clients.groq import call_groq, call_groq_once
 from .factories.payload import GroqPayloadFactory
 from .llm_models import GROQ_MODEL
 from .prompts.qna_prompt import QNA_PROMPT
 
 __all__ = [
-    "call_groq",
-    "call_groq_once",
     "GroqPayloadFactory",
     "GROQ_MODEL",
     "QNA_PROMPT",
