+++ apps/qna/tests/test_chatbots/test_services.py
@@ -44,25 +44,25 @@ class TestInitialService(IsolatedRedisTestClient):
     def test_returns_full_category_path(self) -> None:
         self.assertEqual(self.category, "top > middle > bottom")
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_create_initial_answer_success(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         result = InitialService._create_initial_answer(self.question, self.category)
         self.assertEqual(result, "i am gumba")
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_create_initial_answer_timeout(self, mock: MagicMock) -> None:
         mock.side_effect = GroqTimeoutError
         with self.assertRaises(ExternalAPITimeoutException):
             InitialService._create_initial_answer(self.question, self.category)
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_create_initial_answer_api_error(self, mock: MagicMock) -> None:
         mock.side_effect = GroqAPIError
         with self.assertRaises(ExternalAPIException):
             InitialService._create_initial_answer(self.question, self.category)
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_save_initial_answer_success(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         result = InitialService.save_initial_answer(self.question.id)
@@ -74,7 +74,7 @@ class TestInitialService(IsolatedRedisTestClient):
         self.assertEqual(result.using_model, InitialService.MODEL)
         self.assertIsNotNone(result.created_at)
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_save_initial_answer_raise_409(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         InitialService.save_initial_answer(self.question.id)
@@ -83,7 +83,7 @@ class TestInitialService(IsolatedRedisTestClient):
         self.assertEqual(e.exception.status_code, 409)
         self.assertEqual(str(e.exception), "이미 AI가 답변을 생성했습니다.")
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_save_initial_answer_raise_404(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         with self.assertRaises(NotFoundException) as e:
@@ -91,14 +91,14 @@ class TestInitialService(IsolatedRedisTestClient):
         self.assertEqual(e.exception.status_code, 404)
         self.assertEqual(str(e.exception), "질문 데이터를 찾을 수 없습니다.")
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_save_initial_answer_save_data_equal_cached_data(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         save_data = InitialService.save_initial_answer(self.question.id)
         cached_data = CacheRepository.get_initial(f"qna_initial:{self.question.id}")
         self.assertEqual(save_data, cached_data)
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_get_initial_answer_returns_cached(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         InitialService.save_initial_answer(self.question.id)
@@ -107,7 +107,7 @@ class TestInitialService(IsolatedRedisTestClient):
 
     @patch("apps.qna.services.chatbot_services.sleep")
     @patch("apps.qna.services.chatbot_services.CacheRepository.acquire_lock")
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_get_initial_answer_raises_timeout(
         self, mock_post: MagicMock, mock_lock: MagicMock, mock_sleep: MagicMock
     ) -> None:
@@ -116,7 +116,7 @@ class TestInitialService(IsolatedRedisTestClient):
         with self.assertRaises(GetInitialTimeoutException):
             InitialService.get_initial_answer(self.question.id)
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_get_initial_answer_saves_and_returns(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         result = InitialService.get_initial_answer(self.question.id)
@@ -181,7 +181,7 @@ class TestChatbotService(IsolatedRedisTestClient):
         result = ChatbotService.response_qna_history(self.user_id, self.question_id)
         self.assertEqual(len(result), 2)
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_response_qna_chat_streams_and_saves_history(self, mock: MagicMock) -> None:
         mock.return_value = self.res
         CacheRepository.set_session(
@@ -197,7 +197,7 @@ class TestChatbotService(IsolatedRedisTestClient):
         self.assertEqual(history[0].role, "user")
         self.assertEqual(history[0].content, "질문입니다.")
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_response_qna_chat_raises_timeout(self, mock: MagicMock) -> None:
         mock.side_effect = GroqTimeoutError
         CacheRepository.set_session(
@@ -208,7 +208,7 @@ class TestChatbotService(IsolatedRedisTestClient):
         with self.assertRaises(ExternalAPITimeoutException):
             list(ChatbotService.response_qna_chat(self.user_id, self.question_id, "질문입니다."))
 
-    @patch("apps.qna.chatbot.clients.groq.requests.post")
+    @patch("apps.core.utils.groq_client.requests.post")
     def test_response_qna_chat_raises_api_error(self, mock: MagicMock) -> None:
         mock.side_effect = GroqAPIError
         CacheRepository.set_session(
